% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_blue.R
\name{boxplot_blue}
\alias{boxplot_blue}
\title{Boxplot_blue}
\usage{
boxplot_blue(x, ...)
}
\arguments{
\item{x}{for specifying data from which the boxplots are to be produced. Etiher a numeric vector, or a single list containing such vectors.}

\item{...}{other arguments to be passed to \code{plot} function such as graphical parameters}
}
\description{
\code{boxplot_blue} uses the \code{boxplot} function with default colour blue
}
\examples{

## boxplot on a data frame:
par(las = 1) # all axis labels horizontal
boxplot_blue(mydata, main = "boxplot(*, horizontal = TRUE)", horizontal = TRUE)
}
\seealso{
\code{\link{hist_blue}}, \code{\link{plot_blue}}, \code{\link[base]{boxplot}}
}
