% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_blue.R
\name{hist_blue}
\alias{hist_blue}
\title{Hist_blue}
\usage{
hist_blue(x, ...)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired.}

\item{...}{other arguments to be passed to \code{hist} function such as graphical parameters. See \code{\link[base]{hist}}}
}
\description{
\code{hist_blue} uses the \code{hist} function with default colour blue
}
\examples{
hist_blue(mydata$Norm)
hist_blue(mydata$Uni05)
}
\seealso{
\code{\link{plot_blue}}, \code{\link{boxplot_blue}}, \code{\link[base]{hist}}
}
