% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blue.R
\name{plot_blue}
\alias{plot_blue}
\title{Plot_blue}
\usage{
plot_blue(x, y, ...)
}
\arguments{
\item{x}{the coordinates of points in the plot.}

\item{y}{the y coordinates of points in the plot.}

\item{...}{other arguments to be passed to \code{plot} function such as graphical parameters}
}
\description{
\code{plot_blue} uses the \code{plot} function with default colour blue
}
\examples{
plot_blue(1:4,1:4)
}
\seealso{
\code{\link{hist_blue}}, \code{\link[base]{plot}}
}
